requirements.txt
streamlit
replicate
openai
python-dotenv